requirements.txt
python-pptx>=0.6.21
pytest>=7.0.0
Pillow>=9.0.0
